Add template on ID type to SearchDirectProduct
SearchDirectProduct changed to SearchDirectProductBase, and template
added to allow for changing the type for the ID used by the search.
This will be very useful for EAP, which needs the ID type to be int64_t
instead of the usual int.

SearchDirectProduct is now a typedef that partially specializes
SearchDirectProductBase to int IDs.